// generatedEntities.ts
type Token = {
  id: string
  tokenId: bigint
  collection_id: string
  ownerAddress: string | undefined
  minterAddress: string | undefined
  minter_id: string | undefined
  owner_id: string | undefined
  erc2981_id: string | undefined
  createdBlockNumber: number
  createdTimestamp: number
  updatedBlockNumber: number
  updatedTimestamp: number
}

export const INITIAL_TOKEN: Token = {
  id: "",
  tokenId: 0n,
  collection_id: "",
  ownerAddress: undefined,
  minterAddress: undefined,
  minter_id: undefined,
  owner_id: undefined,
  erc2981_id: undefined,
  createdBlockNumber: 0,
  createdTimestamp: 0,
  updatedBlockNumber: 0,
  updatedTimestamp: 0,
} satisfies Token